Jaewon

+++ frontend/src/components/Project/Project.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useSelector, useDispatch } from "react-redux";
+import {useDispatch, useSelector} from "react-redux";
 import { FiTrash } from "react-icons/fi";
 import axios from "axios";
 import moment from "moment";
@@ -8,17 +8,20 @@ import moment from "moment";
 import UserProfile from "../UI/UserProfile";
 import { projectActions } from "../../store/ProjectInfo";
 import { API } from "../../config";
+import {projectsActions} from "../../store/Projects";
+
 
 const Project = (props) => {
-  const navigate = useNavigate();
 
+  const navigate = useNavigate();
   const dispatch = useDispatch();
-  const user = useSelector((state) => state.userReducer.userObj);
 
   const project_id = props.projectInfo.project_id;
+  const projectInfo = props.projectInfo;
+  const user = props.user;
 
   const [members, setMembers] = useState([]);
-  const [memberIds, setMemberIds] = useState([]);
+
 
   useEffect(() => {
     const memberGetCall = async () => {
@@ -27,16 +30,12 @@ const Project = (props) => {
         const data = res.data;
 
         let memList = [];
-        let memIdList = [];
         for (let idx in data) {
           memList.push(data[idx].username);
-          memIdList.push(data[idx].member_id);
         }
-
         setMembers(memList);
-        setMemberIds(memIdList);
       } catch {
-        alert("초기화 실패 . . .");
+        // alert("Project.js: 초기화 실패 . . .");
       }
     };
     memberGetCall();
@@ -44,23 +43,23 @@ const Project = (props) => {
 
   const onClick = () => {
     dispatch(projectActions.setProject(props.projectInfo));
-
-    navigate(`${project_id}`, {
-      state: {
-        userInfo: user,
-        projectInfo: props.projectInfo,
-        members,
-        memberIds,
-      },
-    });
+    navigate(`${project_id}`);
   };
 
-  const ProjectDeleteClick = async () => {
-    try {
-      await axios.delete(`${API.PROJECT}/${project_id}`);
-      window.location.reload();
-    } catch {
-      alert("프로젝트 삭제에 실패하였습니다.");
+  const ProjectDeleteClick = async (e) => {
+    const title = e.currentTarget.title;
+    if(projectInfo.owner!=user.id){
+      alert("정산을 생성한 유저만 삭제가능합니다.");
+      return;
+    }
+    if(window.confirm( `"${title}"을 삭제하시겠습니까?`)){
+      const res = await axios.delete(`${API.PROJECT}/${project_id}`);
+      if(res.status == 200){
+        dispatch(projectsActions.needUpdate());
+        console.log("정산 삭제 성공.");
+      }else{
+        console.log("정산 삭제에 실패");
+      }
     }
   };
 
@@ -108,12 +107,13 @@ const Project = (props) => {
         </div>
         <hr className="w-full my-1.5 border border-solid border-gray" />
         <div className="flex justify-start mt-2.5 text-xs text-darkgray">
-          정산 마감 기한:{" "}
-          {moment(props.projectInfo.end_dt).format("YYYY-MM-DD")}
+          {/*:{" "}*/}
+          {moment(props.projectInfo.event_dt).format("YYYY-MM-DD")}
         </div>
       </div>
       <button
         className="absolute right-5 bottom-4 text-red"
+        title={props.projectInfo.title}
         onClick={ProjectDeleteClick}
       >
         <FiTrash size="12" />
